Api gateway integration,update of order status after the payment

+++ src/main/java/com/example/orderservice/Repos/orderRepo.java
@@ -1,8 +1,10 @@
 package com.example.orderservice.Repos;
 
 import com.example.orderservice.Models.Order;
+import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
 
-public interface orderRepo extends JpaRepository<Order, Integer> {
+public interface orderRepo extends JpaRepository<Order, Long> {
 
 }
